requirements.txt
requests>=2.31.0
transformers>=4.30.0
python-dotenv>=1.0.0
